Refactor resolve and load config
